requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
black==20.8b1
chardet==3.0.4
click==7.1.2
discord.py==1.5.1
flake8==3.8.4
idna==2.10
mccabe==0.6.1
multidict==4.7.6
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
regex==2020.11.13
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
yarl==1.5.1
